anpi-call-mail/requirements.txt
sendgrid==6.10.0
flask==3.0.0
functions-framework==3.5.0
python-dotenv==1.0.0
